refactor(): hooks folder structure
