+++ src/app/(voting)/reveal/[id]/page.tsx
@@ -16,9 +16,9 @@ import { PaginationDots } from "@/components/dispute-overview/PaginationDots";
 import { SuccessAnimation } from "@/components/SuccessAnimation";
 import { DisputeCandidateCard } from "@/components/disputes/DisputeCandidateCard";
 import { VsBadge } from "@/components/disputes/VsBadge";
-import { useReveal } from "@/hooks/useReveal";
-import { usePageSwipe } from "@/hooks/usePageSwipe";
-import { useDisputeParties } from "@/hooks/useDisputeParties";
+import { useReveal } from "@/hooks/voting/useReveal";
+import { usePageSwipe } from "@/hooks/ui/usePageSwipe";
+import { useDisputeParties } from "@/hooks/disputes/useDisputeParties";
 
 export default function RevealPage() {
   const router = useRouter();
