+++ src/components/dispute-overview/EvidenceView.tsx
@@ -4,8 +4,8 @@ import React, { useRef } from "react";
 import { useRouter } from "next/navigation";
 import { DisputeOverviewHeader } from "@/components/dispute-overview/DisputeOverviewHeader";
 import { PaginationDots } from "@/components/dispute-overview/PaginationDots";
-import { useEvidence, EvidenceRole } from "@/hooks/useEvidence";
-import { usePageSwipe } from "@/hooks/usePageSwipe";
+import { useEvidence, EvidenceRole } from "@/hooks/evidence/useEvidence";
+import { usePageSwipe } from "@/hooks/ui/usePageSwipe";
 import {
   Calendar,
   User,
