+++ src/components/profile/PendingPaymentsDialog.tsx
@@ -11,7 +11,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import { useMyDisputes } from "@/hooks/useMyDisputes";
+import { useMyDisputes } from "@/hooks/disputes/useMyDisputes";
 import { Button } from "@/components/ui/button";
 
 export const PendingPaymentsDialog = () => {
